+++ fakernaija/tests/test_phonenumbers.py
@@ -23,27 +23,29 @@ class TestPhoneNumberProvider(unittest.TestCase):
         """Test that generate_phone_number returns a valid phone number."""
         prefix = "0703"
         phone_number = self.provider.generate_phone_number(prefix)
-        assert phone_number.startswith(prefix)
-        assert len(phone_number) == PHONE_NUMBER_LENGTH
+        self.assertTrue(phone_number.startswith(prefix))
+        self.assertEqual(len(phone_number), PHONE_NUMBER_LENGTH)
 
     def test_phone_number_random(self) -> None:
         """Test that phone_number generates a valid random phone number."""
         phone_number = self.provider.phone_number()
-        assert phone_number.startswith(tuple(self.provider.all_prefixes))
-        assert len(phone_number) == PHONE_NUMBER_LENGTH
+        self.assertTrue(phone_number.startswith(tuple(self.provider.all_prefixes)))
+        self.assertEqual(len(phone_number), PHONE_NUMBER_LENGTH)
 
     def test_phone_number_with_valid_network(self) -> None:
         """Test that phone_number generates a valid phone number for a specific network."""
         phone_number = self.provider.phone_number(network="mtn")
-        assert phone_number.startswith(tuple(self.provider.network_prefixes["mtn"]))
-        assert len(phone_number) == PHONE_NUMBER_LENGTH
+        self.assertTrue(
+            phone_number.startswith(tuple(self.provider.network_prefixes["mtn"])),
+        )
+        self.assertEqual(len(phone_number), PHONE_NUMBER_LENGTH)
 
     def test_phone_number_with_valid_prefix(self) -> None:
         """Test that phone_number generates a valid phone number for a specific prefix."""
         prefix = "0805"
         phone_number = self.provider.phone_number(prefix=prefix)
-        assert phone_number.startswith(prefix)
-        assert len(phone_number) == PHONE_NUMBER_LENGTH
+        self.assertTrue(phone_number.startswith(prefix))
+        self.assertEqual(len(phone_number), PHONE_NUMBER_LENGTH)
 
     def test_phone_number_with_invalid_prefix(self) -> None:
         """Test that phone_number raises ValueError for an invalid prefix."""
@@ -58,14 +60,10 @@ class TestPhoneNumberProvider(unittest.TestCase):
     def test_phone_number_with_network_and_prefix(self) -> None:
         """Test that phone_number generates a valid phone number for a specific network and prefix."""
         phone_number = self.provider.phone_number(network="glo", prefix="0705")
-        assert phone_number.startswith("0705")
-        assert len(phone_number) == PHONE_NUMBER_LENGTH
+        self.assertTrue(phone_number.startswith("0705"))
+        self.assertEqual(len(phone_number), PHONE_NUMBER_LENGTH)
 
     def test_phone_number_with_invalid_network_and_prefix(self) -> None:
         """Test that phone_number raises ValueError for a valid network and invalid prefix combination."""
         with self.assertRaises(ValueError):
             self.provider.phone_number(network="glo", prefix="0703")
-
-
-if __name__ == "__main__":
-    unittest.main()